add dummy record rust code
