+++ rust/aleo/record/mod.rs
@@ -14,68 +14,4 @@
 // You should have received a copy of the GNU General Public License
 // along with the Aleo library. If not, see <https://www.gnu.org/licenses/>.
 
-use snarkvm_algorithms::CRH;
-use snarkvm_dpc::{
-    account::{AccountAddress, AccountPrivateKey},
-    base_dpc::{
-        instantiated::{Components, InstantiatedDPC},
-        record::DPCRecord,
-    },
-    record_payload::RecordPayload,
-    PublicParameters,
-};
-use snarkvm_utilities::{to_bytes, ToBytes};
-
-use rand::{Rng, SeedableRng};
-
-pub fn create_record<R: Rng>(rng: &mut R) -> anyhow::Result<DPCRecord<Components>> {
-    let parameters = PublicParameters::<Components>::load(false).unwrap();
-
-    let spender = AccountPrivateKey::<Components>::new(
-        &parameters.system_parameters.account_signature,
-        &parameters.system_parameters.account_commitment,
-        rng,
-    )?;
-
-    let sn_randomness: [u8; 32] = rng.gen();
-    let old_sn_nonce = parameters.system_parameters.serial_number_nonce.hash(&sn_randomness)?;
-
-    let noop_program_id = to_bytes![
-        parameters
-            .system_parameters
-            .program_verification_key_crh
-            .hash(&to_bytes![parameters.noop_program_snark_parameters.verification_key]?)?
-    ]?;
-
-    let address = AccountAddress::<Components>::from_private_key(
-        parameters.account_signature_parameters(),
-        parameters.account_commitment_parameters(),
-        parameters.account_encryption_parameters(),
-        &spender,
-    )?;
-
-    let record = InstantiatedDPC::generate_record(
-        &parameters.system_parameters,
-        old_sn_nonce,
-        address,
-        true,
-        0,
-        RecordPayload::default(),
-        noop_program_id.clone(),
-        noop_program_id.clone(),
-        rng,
-    )?;
-
-    Ok(record)
-}
-
-#[test]
-fn test_create_record() {
-    use rand_xorshift::XorShiftRng;
-
-    let mut rng = XorShiftRng::seed_from_u64(1231275789u64);
-
-    let record = create_record(&mut rng).unwrap();
-
-    println!("done");
-}
+pub mod dummy;
